feat: rename discussions to threads
