+++ app/bookclubs/events.py
@@ -9,6 +9,6 @@ async def on_user_deleted(payload: dict) -> None:
             payload["user_id"], delete_owned_clubs=payload["delete_clubs"]
         )
         await db.commit()
-    # Треды удалённых клубов чистит discussions - своим событием, не импортом.
+    # Треды удалённых клубов чистит threads - своим событием, не импортом.
     if deleted_club_ids:
         await events.publish(events.CLUBS_DELETED, {"club_ids": deleted_club_ids})
