+++ app/main.py
@@ -3,25 +3,23 @@ from contextlib import asynccontextmanager
 
 from fastapi import Depends, FastAPI, Header, HTTPException, Request, Response, status
 from fastapi.exceptions import RequestValidationError
+from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
-
-from app.core import events
-from app.core.errors.APIException import APIException
-from app.core.models.response_model import ResponseModel
+from fastapi_limiter import FastAPILimiter
+from redis import asyncio as redis_asyncio
 
 # Импорт ради side-effect: регистрация доменных подписчиков на события.
 import app.bookclubs.events  # noqa: F401
-import app.discussions.events  # noqa: F401
-from app.iam.router import auth_router, users_router
+import app.threads.events  # noqa: F401
 from app.bookclubs.router import router as bookclubs_router
 from app.books.router import router as books_router
-from app.discussions.router import threads_router, comments_router
+from app.core import events
+from app.core.errors.APIException import APIException
+from app.core.models.response_model import ResponseModel
 from app.genres.router import router as genres_router
-from fastapi.middleware.cors import CORSMiddleware
+from app.iam.router import auth_router, users_router
 from app.settings import settings
-
-from redis import asyncio as redis_asyncio
-from fastapi_limiter import FastAPILimiter
+from app.threads.router import threads_router, comments_router
 
 
 class UTF8JSONResponse(JSONResponse):
@@ -51,10 +49,10 @@ app = FastAPI(
 if settings.docs_enabled:
     app.router.routes = [r for r in app.router.routes if getattr(r, "path", None) != app.openapi_url]
 
+
     @app.api_route(app.openapi_url, methods=["GET", "HEAD"], include_in_schema=False)
     def openapi():
         return app.openapi()
-
 
 app.include_router(auth_router)
 app.include_router(users_router)
